[COMMON] Allow to search secret by uid even if name is Overridden
In some cases it is useful to bypass default policy for secret name
generation and provide a custom name for a secret. In this case
current search implementation got confused and couln't find a secret
based on uid. This patch fixes the issue by comaring not only name
but also uid.

Issue-ID: OOM-2246
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: Iaea7a23fee09aa388968aad792ba7f7e1fbf2f21