Update patch set 2
Patch Set 2:

Hi Jack,
as you may know, we would like to have to ability to enable or disable https. 
When https is enabled, you should retrieve the https certificates inside AAF.

As far as I understand your files, it seems to be the case (as I see "aaf" stuff) for the AAF part but I don't see the ability to disable it. Can you confirm?

I'm not asking to do the work in this particular patchset (but if it's not big work, it would be nice, "global.aafEnabled being the "well known" toggle), but do you think this would be complex in your side?

Patch-set: 2