Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi Jack,
> as you may know, we would like to have to ability to enable or disable https. 
> When https is enabled, you should retrieve the https certificates inside AAF.
> 
> As far as I understand your files, it seems to be the case (as I see "aaf" stuff) for the AAF part but I don't see the ability to disable it. Can you confirm?
> 
> I'm not asking to do the work in this particular patchset (but if it's not big work, it would be nice, "global.aafEnabled being the "well known" toggle), but do you think this would be complex in your side?

I think this would be possible, but not trivial.  In addition to controlling whether we launch the tls-init-container that fetches the certificates from AAF, we also need to configure each of the components to use TLS or not. Each component has its own way of doing this.  In addition, some components act as clients of other components, and the clients need to know whether to use http or https to access the server components.  Again, each component has its own way of configuring how it accesses its servers.  We also have components that are deployed by Cloudify, via blueprints, sometimes long after the initial deployment of ONAP--controlling their use of HTTPS will require additional work outside of the Helm charts.

Patch-set: 2