Update patch set 2
Patch Set 2: Code-Review+1

@Jack,
Ok, understood.
Actually, I've got an ongoing review that will add a template named "needTLS". With this, we'll know if we want TLS or not.
So, I'll make a review as soon as this is merged.
I propose not to change anything on this review.
We have some troubles on the gating side, I'm on it.

Patch-set: 2