Update patch set 11
Patch Set 11:

> Patch Set 11: Code-Review-1
> 
> (1 comment)
> 
> If you don't mind I'd prefer to split this into smaller chunks (at least 3):
> 1) Update API versions
> 2) Add common templates
> 3) Make DMaaP service Mesh compatible
> 
> It would be way easier to review...


Hi @Krzysztof, Yes I can do that.
As I'm not a gerrit expert, is there a simple way to do that, knowing that the "last" one should be the parent of the rest in list?
or do I have to abandon the "child" patches and recreate new?

And I think I'll break DMaaP part into several pieces (one per component) also if the "reparenting" part is easy

Patch-set: 11