[COMMON] Don't create dummy mysql DB by default in a shared instance
When deploying a shared mariadb-galera instance using common chart
a dummy database is created based on the default values n the chart.

This is obviously unnecessary and creates an obviousl security issue.
That's why let's make sure that when we deploy a shared mariadb
instance no dummy databases are created.

Issue-ID: OOM-2053
Change-Id: I1130cb8eb555b15a2d8b365102d69e32259233eb
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>