Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks, Morgan. I suppose we will learn as we go. :)
> 
> Can you please merge (including parents, though)?

I have the power to merge but as it is a patch in OOM project, I would prefer that someone from OOM does it.
Usually Sylvain is the office next door but he is sunbathing in Tenerife this week. 
If nobody from OOM merges it before tomorrow, I will do.

Patch-set: 1