Update patch set 1
Patch Set 1:

Hey @Vijay, @Jack and @Srini. Here is the initial patch for the k8scloud crd. I would really like your feedback on this, especially for the labels. The labels are not cast in stone and in fact I'm looking for your recommendations as to what labels you require.
Thanks,
Mike.

Patch-set: 1