Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Morgan - AAF comes up fine in the patchset install did in windriver. Not sure why its failing on the gating. Do you have associated defect for AAF on this issue?

@Vijay in the Daily master (same installation than gating) it is also fine (no aaf errors but some oof that could be linked to aaf behavior). It seems that we have some random behavior with AAF. I think it will be the highest risk for frankfurt, this component is key but it is a bit fuzzy today.
 
I can rebase and give a new try on this patch (as it is working on Daily Master, we may expect that it would be also ok for the patch)

In the gating (on Azure or Orange hardware), we reinstall from scratch. On windriver it is not always the case. If somebody made a manual change, use a staging version, it may work but any new installation will fail. That is why we prefer to install from scratch based on the OOM master (no staging) to be sure that it is reproducible.

Patch-set: 3
Label: Verified=0