[NBI] Don't hardcode mariadb-galera password
Let's use common secret template to generate user credentials for NBI
DB and depend on mariadb-galera to generate secure enough root
password.

BTW.
Don't be surprised for now mariadb-galera has a hardcoded root
password but as soon as we move all charts that use it to common
secret template it will be auto generated.

Issue-ID: OOM-2291
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: I5d16f6c26aa63a46db98ba3dab3a76267b4049f1