Update patch set 1
Patch Set 1:

@Morgan
I've tried to debug the issue but to be honest I've no idea what's happening.

Based on the logs I see that DB are created properly and I've not found any logs which would show that the service cannot connect to the DB.

Service is up and running and I can access it with wget.

The only idea that comes to my mind is that robot tests may be trying to access oof using http instead of https as this gave me similar symptoms.

As I'm not a robot specialist could you please take a look and find if robot tries to access cmso using http or https?

Patch-set: 1
Label: Code-Review=0