Update patch set 5
Patch Set 5: Code-Review-1

(107 comments)

Hello,
So first don't be afraid by the number of comments, a lot are very simple to fix.

I really like aafconfig helper and I think it should be given to common/common so we can reuse it on other charts.

on the service part, you may need to create new templates as we have not taken into account yet the ability to create several services for one chart.

on the metadata part, you may need to add the ability to add a role in the template.

thanks!

Patch-set: 5
Label: Code-Review=-1