Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> dcae pods not starting

As with previous patch sets:

(1) This is not a problem with DCAE.  Starting with the Frankfurt release, the DCAE components fetch their TLS certificates from AAF at deployment time.  AAF is not coming up properly in the gating test environment.  This prevents the DCAE components from starting.

(2) This is an environment-specific problem.  AAF comes up (and all DCAE components come up) in the Wind River environment and also in our local test environment (Rancher on a simple KVM VM). I've attached (on more than one occasion) documentation of successful DCAE launches in my local test environment.  In my environment I bring up the subset of ONAP needed to support DCAE (including AAF) from freshly-built charts onto an empty cluster--I'm not resorting to gimmicks to make things work.

The only option within the DCAE team's control is to roll back the TLS changes (so that our deployment is independent of AAF) and return to the highly insecure approach we used for certificates in earlier releases (with "secrets" stored in source control).  The more sensible option, but one which the DCAE team can't implement, is to determine why AAF is failing in the gating environment but not in Wind River and other environments, and then fix the gating environment.

Given that it is late in the Frankfurt release cycle, and given that we have more DCAE-related changes to OOM for Frankfurt, we need to decide very soon what option we are going to take.

I don't know how many other teams are also moving to fetching TLS certs from AAF at deployment time.  (DMaaP is one, but as far as I know they have not submitted any OOM changes for Frankfurt that would be impacted by this problem.) Any other team that attempts this is going to encounter the same problem with AAF in the gating environment.

Patch-set: 5
Label: Code-Review=0