[COMMON / NBI] Handle TLS/Non-TLS for Service
Current service and headlessService templates doesn't handle the fact
that out of cluster ports must be TLS encrypted only.
With a new (backward compatible) DSL, this is now possible.

In values.yaml, all ports in service part with port AND plain_port will
have the ability to be HTTP or HTTPS depending on the context.

Per default, they'll be HTTPS.

An implementation example is done with NBI component.

Issue-ID: SECCOM-256
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: Ibc2f88f5023d5ff4363d86fe48b6b8f9452f9fb6