Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Please use automated certificate generation instead of hardcoding the certificate

The hardcoded certifcates are used by reverse-proxy and forward-proxy, and all the user authentication and authorization are done through AAF. now the CDS-UI is integrated with AAF service for RBAC. 

Automated certifcate generation is done with the init container, you can enable the feature by aaf.enabled flag. The tricky part is to use the auto-generated certifcates/password for each service, revser-proxy, forward-proxy, cds-ui, it involves the certifcate password decryption, obfuscation and configuration. Reverse-proxy and forward-proxy code need some refactoring to support external configuration. cds-ui need to be refactored to support reading certifcate and password as well.

This patch is submitted for the CDS-UI and AAF integation with sidecar implementation. Using the auto-generated certifcate and password would be the next task.

Patch-set: 1