Update patch set 3
Patch Set 3:

(8 comments)

Hello,
I'm not really happy with the jks, keyfile and trust.jks being directly in git.
Can you explain if there's a security breach or not on this point?

Question for information (as we're having a discussion around this): Is it possible to have both HTTP AND HTTPS ports running at the same time on a container or it's only one OR the other? The answer will change the way we create internal services

Patch-set: 3