Update patch set 2
Patch Set 2:

I understand the need but this is not a fair way to proceed to request new features. This PR is about re-introducing the CLAMP dashboard as it was before the searchguard legal issues.
You should open tickets on CLAMP for the features you wish to have and not block this PR because some CLAMP features you want are not there.

Patch-set: 2