Update patch set 3
Patch Set 3: Code-Review-1

@Sylvain
Based on the feedback I got from Ramesh it's not possible to generate client certificate on startup (I still need to confirm this with Amy and AAF team). If this is true then we have 2 options:
1) Put in in container image,
2) Provide it to pod at deployment

Both options are equally insecure and both compromise the certificate but option #2 allows you to at least replace the certificate with your own one at the deployment without rebuilding the containers.

Nevertheless those file should be provided as secrets not as resources so I also vote -1 on this.

Patch-set: 3