Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I understand the need but this is not a fair way to proceed to request new features. This PR is about re-introducing the CLAMP dashboard as it was before the searchguard legal issues.
> You should open tickets on CLAMP for the features you wish to have and not block this PR because some CLAMP features you want are not there.

Hello,

First ask is done by several components so I don't understand the issue no to do it.

Second one would be great if you can do it as Krzysztof would have to do it just after.

And the last one is part of the M3 list (see https://jira.onap.org/browse/OOM-1986) so it's not news no?

Patch-set: 2