Update patch set 7
Patch Set 7: -Code-Review

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > (2 comments)
> > 
> > If I'm not clear, don't hesitate to say it, I'll try to explain another way :)
> 
> It was my intention to change the templatefiles directly in 
> kubernetes/common/common/templates.
> in summary I have extended (recent master from yesterday)
> kubernetes/common/common/templates/_labels.tpl
> kubernetes/common/common/templates/_name.tpl
> kubernetes/common/common/templates/_service.tpl
> kubernetes/common/common/templates/_storage.tpl 
> 
> In the patchset file list all files are visible. 
> Therefore I asked you ;)

Aouch, this one is on me. I was sure it was on elasticsearch folder, sorry :(
And it's actually compatible with other use of the templates, great job!

As gating kind of passed (errors are not there because of your code), I'll make a deeper review ASAP.

Patch-set: 7