[COMMON] Use common secret template in dgbuilder
Taken into account how "easy" it would be to modify the dgbuilder
which is written in JavaScript (which is not my mother tongue to say
the least) let's try to remove hardcoded passwords from config files
without modifying the application container itself.

In order to achieve this:

1) Remove createReleaseDir.sh script from the container as it is never
used and contains a ton of passwords

2) Replace all sensitive values in config files with references to
respective environment variables

3) Introduce init container that will run envsubst command on config
files and copy them from ConfigMap value to the new volume which is
backed by tmpfs so that the plain text passwords are never written to
the disk

For now all the hardcoded values are still there to minimize the risk
of breaking the deployment but step by step they will be removed in
next commits.

Issue-ID: OOM-2247
Change-Id: I5a428e3415713857084ba6aaa6be9b04a8eb8c0f
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>