Update patch set 1
Patch Set 1: Code-Review-1

Not to mention that it would be great to know what is considered as enabled and disabled.

For enable is it enough to define the env variable or it has to be set to some particular string?

When this option is disabled?

And finally why do we need this?

Patch-set: 1
Label: Code-Review=-1