Update patch set 5
Patch Set 5: Code-Review+1

> Patch Set 5: Verified+1
> 
> Build Successful 
> 
> https://jenkins.onap.org/job/oom-master-verify-helm/3682/ : SUCCESS
> 
> Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/oom-master-verify-helm/3682

Krzysztof,
the certificates were originally demo pem certificate, I replace them with the clamp AAF certificates just for the sake of not using the demo one. But if this is really an issue I can as well remove ssl for elasticsearch altogether(and remove those certificates) since it is not actually required here.

I put the ssl verification to false because of fqdn/hostname check on the ssl certificate which would not match. Again I can also just remove the intro of SSL for elasticsearch.

the rest of your remarks have been corrected and the verify build is successful.

Patch-set: 5
Label: Code-Review=+1