Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hello,
> 
> I don't get it @Sylvain :)
> 
> How
> 
>     - name: api
>       port: 8443
>       plain_port: 8080
>       port_protocol: http
>       nodePort: 74
> 
> is supposed to expose 2 nodeports ? ( http and https ) since there is only 74 ?

Hi @Matthieu,
We discussed with krzysztof and Seccom (and decision is not made) but:

* SECCOM doesn't want http nodeports
* internal communication can still be http (but with https preference).

Then we had 2 options:

1/ create one service and depending if we need tls done at container level or done at PaaS level enable or disable https / http port.
2/ create two services: one with NodePort type with https only, one for with ClusterIP with both.

so, as it's not allowed anymore to have http on nodeport, we have created this DSL.

Current templating is doing option 1. (and as SO and AAI cannot have both ports, https and http listening at the same time, I think we'll stick to this one).

Patch-set: 3