Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > this one is conflicting with https://gerrit.onap.org/r/c/oom/+/101684 where I used NBI as an example for service templates "next gen" (sorry @Matthieu ;) )
> > It can be abandoned IMHO
> 
> Hi Sylvain, just to make sure I understand, you are asking Matthieu to abandon this change in this review and use the shared the port AND plain_port instead?. Please confirm as this will have a knock on impact on documentation also. So to switch between http and https for nbi port, how can an external client choose?. We have use case teams using http , so if https by default, they would need to be notified or have some way to use http if needed per deployment.

Hi adrian, as said in the other patch to matthieu:

We discussed with krzysztof and Seccom (and decision is not made) but:

    * SECCOM doesn't want http nodeports
    * internal communication can still be http (but with https preference). 

Then we had 2 options:

1/ create one service and depending if we need tls done at container level or done at PaaS level enable or disable https / http port.
2/ create two services: one with NodePort type with https only, one for with ClusterIP with both.

so, as it's not allowed anymore to have http on nodeport, we have created this DSL.

Current templating is doing option 1. (and as SO and AAI cannot have both ports, https and http listening at the same time, I think we'll stick to this one). 

I propose to continue discussion on https://gerrit.onap.org/r/c/oom/+/101684

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0