[SDNC] Use common secret template in ueb listener
Whole SDNC strongly depends on the assumption that it is using a
common mariadb-galera instance and that root password is secret
password. Also user and password to sdnc DB is hardcoded.

Let's start working on removing this assumption and component by
component add support for local and shared mariadb instance without
hardcoding any passwords to the database.

In this patchAll passwords are still hardcoded in the helm chart to
not break other parts of SDNC. Those values will be removed in a final patch.

Issue-ID: OOM-2309
Change-Id: Ie998b3c5775807ef096074b18a18a1773120c1d6
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>