Update patch set 3
Patch Set 3:

Hi Krzysztof Opasiak, thank you for point out the error. That is indeed a problem,but for the design, one the one hand, we don't want to bound to OOM, and on the other hand, we don't have enough resource to make big change to this design. These changes require sufficient testing to ensure that our original features are not affected.

Patch-set: 3
Label: Code-Review=0