Adjust DelaySeconds parameter to ensure uui-server can start normally
When we switch to https, the startup time is slightly longer. The initial DelaySeconds time that
was originally set is more critical. Sometimes it may cause our program to restart because the
time limit for health check has been reached, so we need to adjust this parameter.
Because it is the modification caused by the https switch change, so I used this issue Id.

Issue-ID: USECASEUI-356
Change-Id: I929814cd586db55f040588dffe2f18832f409c92
Signed-off-by: hekeguang <hekeguang@chinamobile.com>