Update patch set 1
Patch Set 1:

When we switch to https, the startup time is slightly longer. The initial DelaySeconds time that was originally set is more critical. Sometimes it may cause our program to restart because the time limit for health check has been reached, so we need to adjust this parameter. Because it is the modification caused by the https switch change, so I used this issue Id, please help to confirm again whether this commit record is feasible, thank you!

Patch-set: 1
Label: Code-Review=0