Update patch set 3
Patch Set 3: Code-Review+1

It looks like all failures that we have here are a noise.
oof fails because for some reason it get 404 from Music but music pod is up and running and nothing changed there recently

@Morgan
Could you please confirm that failures that we see here are rather unrelated to the content of this patch?

Patch-set: 3
Label: Code-Review=+1