Update patch set 1
Patch Set 1: Code-Review+2

Hmm the gating results seems to be a kind of random for me but in the end of the day at least in one deployment it helped.

As this patch shouldn't be harmful to anything and at it succeeded to fix the issue in recent gating deployment let's merge it and see if the issue is fixed also in other reviews.

Patch-set: 1
Label: Code-Review=+2