Update patch set 3
Patch Set 3: Code-Review-1

So the HV ves is not coming up.

@Vijay
You can find the requested logs here:
https://drive.google.com/file/d/1TEfdDZfWNKl-Bsv3F7I__5a2VwtwmaQG/view?usp=sharing
I've no idea what's happening but the logs from the other review:

https://gerrit.onap.org/r/#/c/oom/+/101036/

in the bootstrap container are the same same. Only diff is in plugin ids and in that review the hv ves collector is visible in the pod list.

Patch-set: 3
Label: Code-Review=-1