Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> This change is ready for review.

if this is a "gating test" to see if new version is OK, we're fine (and we'll not merge it).
If you're OK with the deployment, release a new version and update with release version as only released version are allowed.

But doing deployment tests as such is not forbidden ;)

Patch-set: 2
Label: Verified=0