Update patch set 16
Patch Set 16:

> (1 comment)

Right, you can view this patch so far, as work in progress and a patch on top of El Alto.  It needs to be readjusted for all the OOM Frankfurt changes including the removal of passwords.

Krzysztof, I understood that your team, perhaps Dominik (?) will do the work in the charts that have credentials in the configmaps?   Is that a correct assumption, or that is something that will fall into the individual project teams?

Patch-set: 16