Update patch set 16
Patch Set 16:

> @Jorge
 > So we are working in two folds:
 > 1) Remove passwords that are already in the charts (this is what
 > I'm focusing on currently and Dominic is supporting me with
 > required code changes). Currently I'm working on SO and policy is
 > next on my list (hopefully for tomorrow)
 > 
 > 2) Don't allow new passwords to be introduced in hardcoded form
 > and this is sth that we catch during review and it's up to
 > submitter to fix this but I always offer a help if someone feels
 > lost with the new secret template

Ah, I see, there are newly hardcoded passwords being introduced in this patch, right, so we will have to adapt this patch to play nice with your infrastructure changes.   Thank you both for the comments, we'll address them and touch base with you if we get stuck.

Patch-set: 16
Label: Code-Review=0