Update patch set 4
Patch Set 4:

Dear Krzysztof Opasiak
What you said is very reasonable. Now I want to solve the problem of UUI server reflected in the compilation result. I guess UUI failing refers to the problem that the docker reflected in this URL fails to start.
https://gating-results.onap.eu/results//102416-4/infrastructure-healthcheck/k8s/pod-onap-uui-server-c76554967-xpdkw.html

However, I have a question. I think the version of docker image used here is 2.0.1, but this is a version of last year. It does not contain our latest code. Do you know who is responsible for the configuration? I want to check with him.            
I'm sorry to have bothered you so many times. Thank you very much.

Patch-set: 4