Update patch set 1
Patch Set 1:

Having reviewed a lot of code recently, most if not all the internal components are relying on "http" port of DMaaP...
What I can propose is to use the new (and soon to be merged ;)) service templates on message router (I actually need a "volunteer" to validate) which allow https only on outside and https+http on inside?

If it's fine for you, I can make the review (knowing I'll have to change the sts in order to be compatible with this new template).

What's your thoughts?

Patch-set: 1