Update patch set 1
Patch Set 1:

Sylvain Desbureaux, I am assuming as of today all the ONAP applications are deployed in a single cluster. So  I dont think DCAE or any application have any  need for accessing the DMaaP Message Router service from outside cluster using the NodePort. 

If our goal is to close the HTTP traffic to a service from services that reside inside and outside cluster, this change may not help. If out goal is to close the HTTP traffic originated from outside the cluster we can proceed with this change.

Patch-set: 1