Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> @Jack, the floor is yours now ;) (last error is a missing image for nifi)

It's looking in the wrong repository for the image.  The results show that it's looking for the image at nexus3.onap.org:10001/nifi-registry:0.5.0, but the values.yaml file says it should be apache/nifi-registry:0.5.0.  (See oom/kubernetes/dcaemod/components/dcae-nifi-registry/values.yaml, line 41.)  Is the gating deployment overriding the repository setting?
(We are using Apache's publicly released image for the nifi-registry, so we are sourcing it from the public Docker hub.)

Patch-set: 9