Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> you can also do:
> 
> repository: docker.io
> image: apache/nifi-***
> 
> and it will work directly to docker hub if no global override, or through nexus/whatever if global override

I'd prefer to change nothing, but if I must change, I think I prefer this second approach, because it is explicit about where the image resides.  It does not rely on the repository that's storing ONAP images to also act as a cache for the public docker repository.   In an environment with restricted access to the Internet, if the environment includes a repository that does act as a cache for the public docker repository, the global override could point to it.

Patch-set: 9