Update patch set 2
Patch Set 2:

This time, DCAE came up fully.

Here is my conjecture on what happened during the previous deployment:

The ves collector (like several other DCAE components that are responsible for data collection or analysis) are deployed by Cloudify Manager, not (directly) by Helm, using a plugin that we wrote to deploy into k8s.  The plugin creates a k8s deployment (also often a k8s service as well) and then waits for the container(s) to become ready.  After a configurable interval (default, 30 minutes), if the containers are not ready, the plugin declares a failure and deletes all of the k8s entities (deployment, service, etc) that it created.

I suspect that the plugin actually created the k8s deployment and service for the VES collector but timed out waiting for the VES collector to become ready, and then deleted the deployment and service. I am not sure why the VES collector never became ready, and there's no way to tell without logs.  (And there is no way to be absolutely certain my conjecture is correct.)

Patch-set: 2