Update patch set 2
Patch Set 2:

I'm digging around this dcae issue
interesting thing is that when I tried to delete onap-dcaegen2 I got such error:
Error: Job.batch "onap-dcae-cloudify-manager-map[]-cleanup" is invalid: [metadata.name: Invalid value: "onap-dcae-cloudify-manager-map[]-cleanup": a DNS-1123 subdomain must consist of lower case alphanumeric characters, '-' or '.', and must start and end with an alphanumeric character (e.g. 'example.com', regex used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*'), spec.template.labels: Invalid value: "onap-dcae-cloudify-manager-map[]-cleanup": a valid label must be an empty string or consist of alphanumeric characters, '-', '_' or '.', and must start and end with an alphanumeric character (e.g. 'MyValue',  or 'my_value',  or '12345', regex used for validation is '(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?')

Patch-set: 2