Update patch set 5
Patch Set 5:

Hello, can you help merge this code submission now? 
For the problem of uui-server failure, Sylvain Desbureaux has created a new issue, as follows:
https://jira.onap.org/browse/USECASEUI-404

So the code submitted this time is not the cause of failure, and it succeeds after a new attempt. 
So please help merge the code submitted this time, thank you!

Patch-set: 5