Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Looks like it needs bootstrap-container released. Vijay, Jack, are you planning the release of bootstrap-container in the near future?
> 
> We have a change in to move to bootstrap 1.12.0. (https://gerrit.onap.org/r/c/oom/+/103073)  Not sure what's holding up the merge on that.

Hi Jack, I've made a change in blueprints:
https://gerrit.onap.org/r/c/dcaegen2/platform/blueprints/+/103262
and I need it to be released to make this commit work.
I know that I can't use SNAPHOT version of bootstrap here, I did it just to prove that the problem mentioned by Krzysztof doesn't occur with the latest(snapshot) bootstrap.
https://gating-results.onap.eu/results//103261-2/infrastructure-healthcheck/k8s/pod-onap-dcae-dashboard-9cb5d7f57-46n99.html

Patch-set: 2