Update patch set 2
Patch Set 2:

Joanna put the staging version of Bootstrap container because it contains PRH in a version that has a bugfix. (for a testing reason)

The latest Bootstrap container (1.12.0) does not contain bugfix changes. 

In my opinion, we should release a newer version of the container then apply it to this ticket.

Patch-set: 2
Reviewer: Gerrit User 2807 <2807@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0