Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hello, the issue you submitted about resource over consumption has nothing to do with the code I submitted here. I will submit and verify the relevant code according to Morgan's suggestions later. Thank you very much!

Hey,
I don't really understand what's you're trying to solve here:
our deployments shows that uui server now starts in less than a minute so we're good no?
Where do you have this timing issue?

And putting a very long initial delay is not very good as it makes the whole ONAP wait  a lot.
If you REALLY need this, I would:

set a small (10s?) value for initial
initialDelaySeconds: 10
periodSeconds: 10
successThreshold: 3
failureThreshold: 10

With that, we'll wait 110 seconds but if it appears sooner, we won't have to wait a lot.

Patch-set: 5