CDS SDC listener crashing
The environment specification for AUTO in sdc defines the UEB endpoints
with the namespace appended to message-router "message-router.onap". In
CDS configuration we were only specifying the hostname
"message-router" causing SDC to not find a suitable environments and
returning an error.

Modified the sdc listener configuration in CDS to append the namespace
to the configuration.

Issue-ID: CCSDK-2160

Signed-off-by: Sebastien Premont-Tendland <sebastien.premont@bell.ca>
Change-Id: I7afabb91e43948118490455bdc6ba6a53c28a217