Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > Does anyone from DCAE Team knows the reason why dashboard is failing?
> 
> Not seeing this in my local environment.  There were no changes related to dashboard that I know about.  Will forward the log output to a dashboard developer.

The problem is that the dashboard stores a password value into an XML attribute.  The password is obtained dynamically at deployment time from AAF.  It turns out that occasionally the password may contain a character (such as "&") that is significant to XML, potentially causing a parsing error such as the one seen here.   The problem will occur only intermittently--the password is generated on the fly by AAF, so it is different on each deployment. DCAEGEN2-2139 is tracking this.

Patch-set: 3
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0