Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> @Jorge
> So we are working in two folds:
> 1) Remove passwords that are already in the charts (this is what I'm focusing on currently and Dominic is supporting me with required code changes). Currently I'm working on SO and policy is next on my list (hopefully for tomorrow)
> 
> 2) Don't allow new passwords to be introduced in hardcoded form
> and this is sth that we catch during review and it's up to submitter to fix this but I always offer a help if someone feels lost with the new secret template

Krzysztof, we are assuming then that when you are done with SO you will patch this set of charts with the necessary changes to support password removal, is that correct? I don't think that our team will be able to guess what changes will be necessary.

Patch-set: 16
Label: Code-Review=0