Update patch set 3
Patch Set 3:

Krzysztof, I was NOT aware dynamic cert is onap must-have for Frankfurt release. Do you have an epic that is tracking this feature? We will need to apply for TSC exception. I asked Raj to complete this patch using static certs in frankfurt and for onap G release our plan is to enable dynamic certs for CDS UI.

Patch-set: 3