Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Sylvian, it seems the in the example you shared secret is created for the env variables. Is it also supports the binary files. If yes, can you share the correct example

the env variable is substituted at run (via envsubst command that we see in deployment) and thus the "real" configuration file (which is stored in memory in order to be safer) doesn't have the env variable but the interpolated value (which is stored into a secret).

Let's wait gating to see if everything is OK and I'll try to make a proposition (@Krzyzstof has done all this stuff as of today but I believe we should be at least two to be able to do that).

Patch-set: 2