Update patch set 3
Patch Set 3:

ok. I think its a generic concern for anyone that trying to connect to a pod (created by statefulset and hostname changed by headless service)directly. May be we should capture the cluster domain, if we are anticipating someone change the default cluster domain

You are working on using the common secret template for creating the secret  right

Patch-set: 3
Label: Code-Review=0