Update patch set 3
Patch Set 3:

Oh I thought it's related to some OJSI ticket if not then forget about waiver for OJSI;)

So what's the main motivation for doing this? Is it just because you prefer to be safe (which is great if that's the case!) or this is required by some use case or sth?

Patch-set: 3