Add VFC env about reg_to_msb_when_start to determine whether register to microserice
Becaue now oom can register the microservice to msb automatically,
if without K8S, we can also register vfc microservice to msb
automatically by vfc container itself.

IF it is set to false, vfc contanier will not register again, if it is
set to true, vfc will register by itself.
we use this flag to determine who is responbile for serice registeration
and it can reduce duplicate registration.

Change-Id: Ie79a8803725ae0c762d9c39a04f922d9823c46e2
Issue-ID: VFC-1601
Signed-off-by: yangyan <yangyanyj@chinamobile.com>