Update patch set 2
Patch Set 2:

I left the deprecated properties in the configMap on purpose to be informative about change.  I don't agree that it is clutter.  Rather, they help a user more quickly determine what might be different in their config. So I'd like to keep the comments for 1 Release, and then add a Jira to remove them in next Release.  OK?

Patch-set: 2