Update patch set 20
Patch Set 20:

Folks, took the review out of WIP, for discussion purposes, there are 2 missing pieces: 1) the password removal work, and 2) update to released images (we haven't released yet).

How we should be proceeding? I understood that for 1) the password work would be contributed by perhaps Krzystof's team, and 2) we still need to do it ourselves, probably would make sense to do 1) before 2).

I just want to start discussion and get clarity on the logistics on how we are going to get the work done.

Feedback?

Patch-set: 20