[COMMON] Optimize common secret template
It turned out that our current implementation of common secret
template is really heavy which makes onap linitng extremely long.

To improve the situation let's introduce some results caching instead
of processing templates over and over.

For now we cannot simply replace common secret template because in
mariadb-init we generate list of secrets on the fly so we will need
to revisit this fragment later.

Whole series of patches managed to reduce ONAP linting time to 40
mins.

Issue-ID: OOM-2051
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: Id2e743147afa37290df19b73feee67621f13f67c