Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> we have 4 unavailble DCAE component: {"type":"summary","count":11,"ready":7,"items":[{"name":"onap-dcae-cloudify-manager","ready":1,"unavailable":0},{"name":"onap-dcae-config-binding-service","ready":1,"unavailable":0},{"name":"onap-dcae-inventory-api","ready":1,"unavailable":0},{"name":"onap-dcae-servicechange-handler","ready":1,"unavailable":0},{"name":"onap-dcae-deployment-handler","ready":1,"unavailable":0},{"name":"onap-dcae-policy-handler","ready":1,"unavailable":0},{"name":"onap-dcae-dashboard","ready":1,"unavailable":0},{"name":"dep-dcae-ves-collector","ready":0,"unavailable":1},{"name":"dep-dcae-tca-analytics","ready":0,"unavailable":1},{"name":"dep-dcae-prh","ready":0,"unavailable":1},{"name":"dep-dcae-hv-ves-collector","ready":0,"unavailable":1}]}
> 
> is it normal?

All of the DCAE components that are deployed with Helm charts came up successfully, but none of the "service components" that are launched using Cloudify Manager came up.   Unfortunately, the  logs for the dcae-bootstrap container are truncated, so we cannot see why the components were not launched. There is enough in the logs to show that Cloudify Manager was healthy and that the boostrap container could send commands to it, but nothing about the attempts to launch the service components.

This is unlikely to be a problem with the Helm charts, but it is not what we expect.

Patch-set: 3